requirements.txt
gunicorn==20.0.4
Flask==1.1.2
Flask_Cors==3.0.10
pypugjs==5.9.8
pretty_errors==1.2.19
psycopg2==2.8.6
libsass==0.20.1
pypng==0.0.20
